requirements.txt
pyrogram>=2.0.0
tgcrypto>=1.2.5
motor>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
